Remove uint16_t casts before integral promotion
The two operands of the | operator are promoted to int, so the uint16_t
casts are misleading and could be omitted.

Note: We could, however, cast the result of the | operator to uint16_t.

+++ examples/av1_enc_fuzzer.cc
@@ -70,8 +70,7 @@ uint8_t ReadU8(FuzzReader *reader) {
 
 uint16_t ReadU16(FuzzReader *reader) {
   if (reader->size < 2) return 0;
-  const uint16_t value = static_cast<uint16_t>(reader->data[0]) |
-                         static_cast<uint16_t>(reader->data[1] << 8);
+  const uint16_t value = reader->data[0] | (reader->data[1] << 8);
   reader->data += 2;
   reader->size -= 2;
   return value;
